source/requirements.txt
openai
langchain
langchain_core
langchain_community
langchain_openai
pinecone-client
python-dotenv
tiktoken
pypdf
langchain-elasticsearch
elasticsearch
streamlit
unstructured-client
